feat: rename "command" into "builder"
So it's less confusing what this file does, it only contains the
builder parts to create commands
