+++ src/commands/referral/referralRandom.ts
@@ -1,7 +1,7 @@
 import { SlashCommandSubcommandBuilder } from 'discord.js';
 import { getPrismaClient } from '../../clients';
 import { getRandomIntInclusive } from '../../utils';
-import { AutocompleteHandler, CommandHandler } from '../command';
+import { AutocompleteHandler, CommandHandler } from '../builder';
 import { cleanupExpiredCode } from './cleanupExpiredCode';
 import { searchServices } from './services';
