+++ src/commands/reminder/update.ts
@@ -1,5 +1,5 @@
 import { SlashCommandSubcommandBuilder } from 'discord.js';
-import { CommandHandler, Subcommand } from '../command';
+import { CommandHandler, Subcommand } from '../builder';
 import { updateReminder } from './reminder-utils';
 import { convertDateToEpoch } from '../../utils/dateUtils';
