+++ src/commands/serverSettings/set-reminder-channel.ts
@@ -1,5 +1,5 @@
 import { SlashCommandSubcommandBuilder } from 'discord.js';
-import { CommandHandler, Subcommand } from '../command';
+import { CommandHandler, Subcommand } from '../builder';
 import { setReminderChannel } from './server-utils';
 
 export const data = new SlashCommandSubcommandBuilder()
