+++ src/index.ts
@@ -5,7 +5,7 @@ import { processMessage } from './utils';
 import { getDiscordClient } from './clients';
 import { getConfigs } from './config';
 import { commandList, contextMenuCommandList } from './commands';
-import { deployGlobalCommands } from './commands/command';
+import { deployGlobalCommands } from './commands/builder';
 
 const env = dotenv.config();
 dotenvExpand.expand(env);
